cloud_functions/merge_weekly/requirements.txt
google-cloud-storage==2.8.0
google-cloud-firestore==2.10.4
pytz==2023.3